LayoutTests/http/tests/frames/frames-with-same-name-cross-origin-expected.txt

Make sure that frame lookup by name has correct ordering when duplicate names are involved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crossOriginWindow.foo === crossOriginWindow[0] is true
PASS open('', 'foo') === crossOriginWindow.foo is true
PASS successfullyParsed is true

TEST COMPLETE
